Polishin saving structures and variable formations, added debug testing
- reduced memory usage by including sample array in temporal array
- reworked saving structure, it now is separated into the three parts
for save_computational, save_zoom and save_sample
- reworked sampling and zooming functions, memory usage was reworked as
well as general structure checked
- zoom and sample variables now reuse the temporal array to minimize the
usage
- temporal array size was rechecked and documented
- added debugging options to compute global quantities for verbose=4
- changed python testing to include debug testing, added most
fundamental test (not completely verified yet)

+++ Cuda-CMM/src/grid/cmm-grid2d.h
@@ -26,7 +26,7 @@ struct TCudaGrid2D
 {
 	public:
 		int NX, NY, NZ, NX_fft;
-		long long int N, Nfft;
+		size_t N, Nfft;
 
 		double hx, hy, hz;
 		double bounds[6];
